Fix siteId parameter type to match schema

+++ src/main/resources/crafter/studio/database/upgrade/5.0.x/4.x-to-5.x/4.x-to-5.0.new-schema.sql
@@ -147,7 +147,7 @@ END ;
 
 /***************** TEMPORARY PROCEDURE UPDATES *****************/
 CREATE PROCEDURE populateItemTarget(
-	IN siteId INT,
+	IN siteId BIGINT,
 	IN publishTarget VARCHAR(20),
 	IN publishedLastCommit VARCHAR(40))
 BEGIN
